model for external object interpolation
Here the main difference between this and our previous test case is the fact that we don’t pass our data directly into the interpolation function. Instead, we create a special variable vector whose type is VectorTable->for now think of it as something that represents (only) our interpolation data. Other than the creation of the vector object, the rest of the model is virtually identical to the previous case except that we use the InterpolateExternalVector function to perform our interpolation and we pass the vector variable into that function in place of our raw interpolation data.